nquads backup is predictable
We were receiving nightly PRs because the backup of the database was being entirely rewritten, randomly, on a nightly basis. This fix makes the database backup predictable (i.e. sorted) and therefore it only changes when there is actually a change to the database. PRs should reflect actual changes to the DB from now on.